remove password encode logic and serialnumber config from principal config of hub module; mqtt client must send username for topic authentication, even if two-way cerfificate authentication is used
