+++ example/docker/var/db/openedge/application.yml
@@ -5,12 +5,16 @@ services:
     replica: 1
     ports:
       - 1883:1883
+      - 8883:8883
     mounts:
       - name: localhub-conf
         path: etc/openedge
         readonly: true
+      - name: localhub-cert
+        path: var/db/openedge/cert
+        readonly: true
       - name: localhub-data
-        path: var/db/openedge
+        path: var/db/openedge/data
       - name: localhub-log
         path: var/log/openedge
   # - name: agent
@@ -84,6 +88,8 @@ volumes:
     path: var/db/openedge/localhub-conf
   - name: localhub-data
     path: var/db/openedge/localhub-data
+  - name: localhub-cert
+    path: var/db/openedge/localhub-cert-only-for-test
   - name: localhub-log
     path: var/db/openedge/localhub-log
   # agent
