+++ example/docker/var/db/openedge/localhub-cert-only-for-test/server.key
@@ -1,0 +1,27 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIEpAIBAAKCAQEAtm5mCoAriGIrW+Bg5e9zuiTcoCztkfK5MuFlHZLArFK+VsbO
+kZal+v3hvD5uxrDgfX1tToDssJYpgILCZvy3HZ+V//YDm/7FoAqScohk1JGuz6E1
+JC+Puulgv8t3qqNsbM92l92wyrMcNs2W5ObdxeqfIDOFYnC56p4/mb9jjXg9HLm7
+T3itpAQZk0LPau5TcXwLhUiIKm8JlCsXaT6xWTEZCW1RXjOFiQC6iQZp86DKPcCA
+ukXXROjODqZkRtvLGYuTTxuk2yAqhrihPjouswBP7NWJMw/d12d6eG0LHCGe/fxv
+sMNrtuZH3Pa5mPzaICPAqTEiwVbBB8p8eNFWuQIDAQABAoIBABevm8YPQBe0D0cd
+idcw0MjBUnflfW7uoccOZq+BR0zYbWgc5eMXOTOuVOm4/W8EthD1AQUpThHbXUYX
+NCBk/W12M+aIzY0qdgRCeyMtvFGxHI5OsvSgzDuLY4ljZIUeNMyn3xD2sqaOvMIM
+zWOo6fBxdvmaVt4iHrFsnHtc0xMh6jbIIuNVyrG/JeT0bEZehoJQN3QrEgNHpbU/
+0urFxGFOtVwIJcHUDER1Gwms7GVk6ngYsnzrt6DbHVmd8louo69Nekb8VwMWumYl
+Gkw+DmfyrP7nfip7gvuu8eSDGfVI8FQjRaVvDqmBZWoxKYlD5kEn+OLKCMg1Xf2s
+BXo6+lECgYEA4rayW6ccZIALpVHAPzgq/zkbkOHc/48EAC2oL3qbTVBfhoneJTfw
+NXtGtZYRc3eQsOe+fgBbriBRinvP4inBWfzTAf5mJ5II1mkbg7RNqeCX83bcfzkA
+NuVMM5D0hDVAwc4lykcJNTkdvJd3FYUY2AcRTJseDTbhH4j/i1kXyqUCgYEAzf9O
+LOs4g0e3mZball9EEdsWzlNQ30YTL8NPXTBLgW0P1GyPVEjeEjZNGAev6e3l7Dyr
+rUtIdpEHHeJYgofwLOEduoyH8IgCIi1VYKZYJXMnUUBgFKhytewui20Flchf3iW2
+KPT5LfXLOVw54VHLSsiV9QKGJSXohcFILz5Wo4UCgYEAziM6FxXiPgsjC6VCfPkQ
+N1Hwm5KShBfAlnztEIaVkP3EbasLvqGXGrnFODSEF44elIuSbDTdJ30hRx5+jXBT
+6wNzxLtQnSguOu4E44Lcntqkru1jXZYxeLiSycxCV7en8tADEmH2oFsXi7PYxktn
+pH0AoLvl3jQJC8hgOo6+kDUCgYA1/1VjqujuaK804Hk974ztTUvGG8rYLAs9eunj
+2JQCQi9jFG3hvxQPHBLLWMCBBm49pY2u6Ev/YHbIG5vt42pOqVfcnZfIsi40/UMj
+CZeTgYLt5n0JDTq8e0OcIqDTKWoP1jzQsZ5bsNFroMQ/TuyGFtxQFDBLrosr/IVK
+dQHK7QKBgQCFr3gUKAoRiwmWAzU99Ofdxuo600ooARjYJ/+9RzJY+nvVmsA0oy9L
+uiXnHnmsl89Q6+LdC4XSqhJ4UwK3d/ApnOmqS3Xn1Xcr2KJhr+AU3YDigMKBDm9X
+3sRVewNfQ7lYPgXVFcWKDsTX61H+KHIbRrim66tirgCVloKHsaGYpA==
+-----END RSA PRIVATE KEY-----
