+++ example/native/var/db/openedge/application.yml
@@ -7,8 +7,11 @@ services:
       - name: localhub-conf
         path: etc/openedge
         readonly: true
+      - name: localhub-cert
+        path: var/db/openedge/cert
+        readonly: true
       - name: localhub-data
-        path: var/db/openedge
+        path: var/db/openedge/data
       - name: localhub-log
         path: var/log/openedge
       # native process mode only
@@ -93,6 +96,8 @@ volumes:
     path: var/db/openedge/localhub-conf
   - name: localhub-data
     path: var/db/openedge/localhub-data
+  - name: localhub-cert
+    path: var/db/openedge/localhub-cert-only-for-test
   - name: localhub-log
     path: var/db/openedge/localhub-log
   # agent
