+++ example/native/var/db/openedge/localhub-cert-only-for-test/ca.pem
@@ -1,0 +1,25 @@
+-----BEGIN CERTIFICATE-----
+MIIEOzCCAyOgAwIBAgIIcX3314U93DYwDQYJKoZIhvcNAQELBQAwcDELMAkGA1UE
+BhMCQ04xGjAYBgNVBAMMEWlvdGVkZ2UuYmFpZHUuY29tMQ4wDAYDVQQKDAVCYWlk
+dTEMMAoGA1UECwwDQkNFMScwJQYJKoZIhvcNAQkBFhh3YW5neGlhb2NoZW4wNEBi
+YWlkdS5jb20wHhcNMTgwOTI3MTExNDI3WhcNMzgwOTIyMTExNDI3WjBwMQswCQYD
+VQQGEwJDTjEaMBgGA1UEAwwRaW90ZWRnZS5iYWlkdS5jb20xDjAMBgNVBAoMBUJh
+aWR1MQwwCgYDVQQLDANCQ0UxJzAlBgkqhkiG9w0BCQEWGHdhbmd4aWFvY2hlbjA0
+QGJhaWR1LmNvbTCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBALOeir0+
+BlrY1gu6oFu8mjkhwugv1LQ8SD6PSn/nO3J8iTflr1JHTtpcf3XlWCjs5DxTrHyE
+BWpWA1XHSlI1JBlNG/BeVGW/eeYtNb1XHqouhHRbRJI9y3uAE++mdqOyuCRH1WKX
+WMKz0/C+BSUO1i3M6iWMBGccEcPoqFmFZnQ7SRTGqw2W1t8/oZPoXhUn4UeoLw4j
+Gba8bpeH31IXLMtmQ53QSrBSdCcRdw3qXi5+dg3rHKmphao5y3ig+s6WRvyI67Wt
+g+2C0EWLujwLs9Z+F+HzHcEi/3IP1P7KD1InHVAXtiD9J3hnry6+6aUwpcMh1CvC
+4tuSrgCHjyVLjsUCAwEAAaOB2DCB1TAdBgNVHQ4EFgQUI0xfQazD911ZkjAzTojM
+vRtSkBkwDwYDVR0TAQH/BAUwAwEB/zAfBgNVHSMEGDAWgBQjTF9BrMP3XVmSMDNO
+iMy9G1KQGTA6BgNVHSAEMzAxMC8GA1UdHzAoMCYGCCsGAQUFBwIBFhpodHRwOi8v
+cGtpaW92LmJhaWR1YmNlLmNvbTA2BggrBgEFBQcBAQQqMCgwJgYIKwYBBQUHMAGG
+Gmh0dHA6Ly9wa2lpb3YuYmFpZHViY2UuY29tMA4GA1UdDwEB/wQEAwIBxjANBgkq
+hkiG9w0BAQsFAAOCAQEAHkiKSciUuY5L9b86JWFc+wzYDrbE3daSVH3KwactDkwZ
+ocsG8dvJ6ma2RQX08numGLKUxcNrzU5A4wN0SyrcmQ2Jeu6OB363s5xleppRING1
+KsC9v8Mft4BJ4cGQ9ybffYuEPye+ne+LUZQyI5CfWxLvz+MqCqj19fdU3er0m3QY
+VwXw8+Nk8FT+2HrKHfP+4U97L7wWJN4C5/SLJdrFDcmsqQETUuf3l3mLwoCqxDOD
+468yK2cV26WYSFZD7+3Ok+9t1JmMpMwLI8Uq5Wp1UyEBIgQLCdodhzCj+7gCEwWQ
+P0zQbDZBaMViCWXMyJTMCt+BSDXiFC55qaXY4VrRkQ==
+-----END CERTIFICATE-----
