+++ example/native/var/db/openedge/localhub-cert-only-for-test/server.pem
@@ -1,0 +1,26 @@
+-----BEGIN CERTIFICATE-----
+MIIEXDCCA0SgAwIBAgIIQUyqrH0y3rQwDQYJKoZIhvcNAQELBQAwcDELMAkGA1UE
+BhMCQ04xGjAYBgNVBAMMEWlvdGVkZ2UuYmFpZHUuY29tMQ4wDAYDVQQKDAVCYWlk
+dTEMMAoGA1UECwwDQkNFMScwJQYJKoZIhvcNAQkBFhh3YW5neGlhb2NoZW4wNEBi
+YWlkdS5jb20wHhcNMTgxMTAyMDkyMTU0WhcNMzgxMDI4MDkyMTU0WjCBjTFgMF4G
+A1UEAxNXYzNjNGY0MDAyZWE4NDM3NmJhMmJkNDlhY2EyMTg1YjIuYjVkZTE0YTBm
+ZmE2NDhmOWE2NjJhYjNmMTA2ZWRkMTcuaW90LmVkZ2UuYmFpZHViY2UuY29tMQsw
+CQYDVQQGEwJDTjEOMAwGA1UEChMFQmFpZHUxDDAKBgNVBAsTA0JDRTCCASIwDQYJ
+KoZIhvcNAQEBBQADggEPADCCAQoCggEBALZuZgqAK4hiK1vgYOXvc7ok3KAs7ZHy
+uTLhZR2SwKxSvlbGzpGWpfr94bw+bsaw4H19bU6A7LCWKYCCwmb8tx2flf/2A5v+
+xaAKknKIZNSRrs+hNSQvj7rpYL/Ld6qjbGzPdpfdsMqzHDbNluTm3cXqnyAzhWJw
+ueqeP5m/Y414PRy5u094raQEGZNCz2ruU3F8C4VIiCpvCZQrF2k+sVkxGQltUV4z
+hYkAuokGafOgyj3AgLpF10Tozg6mZEbbyxmLk08bpNsgKoa4oT46LrMAT+zViTMP
+3ddnenhtCxwhnv38b7DDa7bmR9z2uZj82iAjwKkxIsFWwQfKfHjRVrkCAwEAAaOB
+2zCB2DAdBgNVHQ4EFgQU/jFfBU/JWUeMghpOYosM3aZpQREwDAYDVR0TAQH/BAIw
+ADAfBgNVHSMEGDAWgBQjTF9BrMP3XVmSMDNOiMy9G1KQGTArBgNVHR8EJDAiMCCg
+HqAchhpodHRwOi8vcGtpaW92LmJhaWR1YmNlLmNvbTA2BggrBgEFBQcBAQQqMCgw
+JgYIKwYBBQUHMAGGGmh0dHA6Ly9wa2lpb3YuYmFpZHViY2UuY29tMA4GA1UdDwEB
+/wQEAwID+DATBgNVHSUEDDAKBggrBgEFBQcDATANBgkqhkiG9w0BAQsFAAOCAQEA
+CWga2HpHvU9pzf7GyH8fFFbFxxMm8QLoPUydB1pnvAD/wLkJE0bVki6i+ysWFUHh
+izVvmlbCJd93cxpAX4oAmbzMb8OOfsNmOCZyS6lq/Qg4BdsyMrfQZMN4HzdqskvK
+OL+8KBdcF9ODnenzqpHhUUuDhM8VJGPl1zTTQ8wEJc8VAlKjxZwZRpNqrER5Yzoi
+dpWwp+Hu4va+sR3N5ulnBYmV9PrZPEMWD0El77ocQcmESCLR1daoO9kbJZqc6uVK
+K8XZYqGnrxbumokwRiqw1Aeat0anloOXanZkqTd144o1QLgIwBhvSuMSpWE6vmaj
+xVSSb5uc79GE7Vj91EDfDQ==
+-----END CERTIFICATE-----
