+++ openedge-hub/config/auth.go
@@ -8,8 +8,7 @@ type Permission struct {
 
 // Principal Principal
 type Principal struct {
-	Username     string       `yaml:"username" json:"username"`
-	Password     string       `yaml:"password" json:"password"`
-	SerialNumber string       `yaml:"serialnumber" json:"serialnumber"`
-	Permissions  []Permission `yaml:"permissions" json:"permissions"`
+	Username    string       `yaml:"username" json:"username"`
+	Password    string       `yaml:"password" json:"password"`
+	Permissions []Permission `yaml:"permissions" json:"permissions"`
 }
